#1363 move XADataSourceMapConverter & XATransactionManagerSPILoader => transaction.xa package.
