+++ sharding-proxy/src/main/java/io/shardingsphere/shardingproxy/backend/jdbc/JDBCBackendHandler.java
@@ -48,7 +48,7 @@ import io.shardingsphere.shardingproxy.transport.mysql.packet.command.query.Quer
 import io.shardingsphere.shardingproxy.transport.mysql.packet.generic.EofPacket;
 import io.shardingsphere.shardingproxy.transport.mysql.packet.generic.ErrPacket;
 import io.shardingsphere.shardingproxy.transport.mysql.packet.generic.OKPacket;
-import io.shardingsphere.transaction.xa.manager.XATransactionManagerSPILoader;
+import io.shardingsphere.transaction.xa.XATransactionManagerSPILoader;
 import lombok.RequiredArgsConstructor;
 
 import javax.transaction.Status;
