+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/manager/XADataSourceMapConverterTest.java
@@ -19,6 +19,7 @@ package io.shardingsphere.transaction.xa.manager;
 
 import com.atomikos.jdbc.AtomikosDataSourceBean;
 import io.shardingsphere.core.constant.DatabaseType;
+import io.shardingsphere.transaction.xa.XADataSourceMapConverter;
 import org.apache.tomcat.dbcp.dbcp2.BasicDataSource;
 import org.junit.Test;
