+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/manager/XATransactionManagerSPILoaderTest.java
@@ -17,6 +17,7 @@
 
 package io.shardingsphere.transaction.xa.manager;
 
+import io.shardingsphere.transaction.xa.XATransactionManagerSPILoader;
 import io.shardingsphere.transaction.xa.manager.atomikos.AtomikosTransactionManager;
 import org.junit.Test;
